requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-chroma>=0.0.1
langchain-core>=0.1.0
openai>=1.3.0
python-docx>=0.8.11
pymupdf>=1.22.5
watchdog>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
chromadb>=0.4.18
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6 